requirements.txt
# LiveKit Agents Framework
livekit-agents>=0.8.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-cartesia>=0.4.0
livekit-plugins-openai>=0.8.0
livekit-plugins-silero>=0.6.0
livekit-plugins-bey>=0.2.0

# Database
supabase>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0

# Date/time parsing
python-dateutil>=2.8.0

# Token server
fastapi>=0.109.0
uvicorn>=0.27.0
livekit-api>=0.6.0
